+++ tests/share/test_secretsmanager.py
@@ -93,7 +93,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -115,7 +115,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
             with self.assertRaisesRegex(
                 ValueError, "Must be provided region in arn: arn:aws:secretsmanager::123456789:secret:plain_secret"
@@ -134,7 +134,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
             with self.assertRaisesRegex(
                 ValueError,
@@ -154,7 +154,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
             with self.assertRaisesRegex(
                 ValueError,
@@ -175,7 +175,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -196,7 +196,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaises(ClientError):
@@ -213,7 +213,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -233,7 +233,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -254,7 +254,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -276,7 +276,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -297,7 +297,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -318,7 +318,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             with self.assertRaisesRegex(
@@ -341,7 +341,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secrets_manager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
             mock_fetched_data = aws_sm_expander(config_yaml)
 
@@ -355,7 +355,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secrets_manager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "mock_elastic_username"
                             password: "mock_elastic_password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             assert mock_fetched_data == parsed_config_yaml
@@ -372,7 +372,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                             username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                             password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
             mock_fetched_data = aws_sm_expander(config_yaml)
 
@@ -386,7 +386,7 @@ class TestAWSSecretsManager(TestCase):
                             elasticsearch_url: "mock_elastic_url"
                             username: "mock_elastic_username"
                             password: "mock_elastic_password"
-                            datastream: "datastream"
+                            es_datastream_name: "es_datastream_name"
             """
 
             assert mock_fetched_data == parsed_config_yaml
@@ -402,7 +402,7 @@ class TestAWSSecretsManager(TestCase):
                         elasticsearch_url: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:url"
                         username: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:username"
                         password: "arn:aws:secretsmanager:eu-central-1:123456789:secret:es_secrets:password"
-                        datastream: "arn:aws:secretsmanager:eu-west-1:123456789:secret:binary_secret"
+                        es_datastream_name: "arn:aws:secretsmanager:eu-west-1:123456789:secret:binary_secret"
             """
             mock_fetched_data = aws_sm_expander(config_yaml)
 
@@ -416,7 +416,7 @@ class TestAWSSecretsManager(TestCase):
                         elasticsearch_url: "mock_elastic_url"
                         username: "mock_elastic_username"
                         password: "mock_elastic_password"
-                        datastream: "mock_nginx.log"
+                        es_datastream_name: "mock_nginx.log"
             """
 
             assert mock_fetched_data == parsed_config_yaml