+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ##### Features
 * Add support for AWS IaC deployment with CloudFormation and terraform: [#115](https://github.com/elastic/elastic-serverless-forwarder/pull/115)
 ##### Breaking changes
-* Replace `es_index_or_datastream_name` config param with `datastream` one: [#115](https://github.com/elastic/elastic-serverless-forwarder/pull/115)
+* Replace `es_index_or_datastream_name` config param with `es_datastream_name` one: [#115](https://github.com/elastic/elastic-serverless-forwarder/pull/115)
 
 ### v0.29.1 - 2022/05/04
 ##### Bug fixes